requirements.txt
discord.py>=2.3.0
openai>=1.6.0
python-dotenv>=1.0.0
